Rework labels for all multi-sts & multi-deployment apps
Any app that uses more than one statefulset or deployment should now be labeled with app: for the actual app the statefulset/deployment is, and component: for the app it is part of.

This makes things a bit neater when there's mariadb, postgres, redis, or rabbitmq deployed alongside an app that's needed for that app to work.